requirements.txt
asgiref                       ==3.8.1
certifi                       ==2024.8.30
charset-normalizer            ==3.4.0
cloudinary                    ==1.41.0
Django                        ==5.1.4
django-cors-headers           ==4.6.0
django-filter                 ==24.3
django-jazzmin                
djangorestframework           ==3.15.2
djangorestframework-simplejwt ==5.3.1
idna                          ==3.10
pip                           ==23.2.1
psycopg                       ==3.2.3
psycopg-binary                ==3.2.3
psycopg2-binary               ==2.9.10
PyJWT                         ==2.10.1
requests                      ==2.32.3
six                           ==1.17.0
sqlparse                      ==0.5.3
stripe                        ==11.3.0
typing_extensions             ==4.12.2